+++ tests/misc/fallback_test/src/071_isset.php
@@ -1,0 +1,5 @@
+<?php
+// https://github.com/phan/phan/issues/3767
+function test71() {
+    return isset ? 2 : 3;
+}